+++ packages/at_commons/test/syntax_test.dart
@@ -299,6 +299,23 @@ void main() {
               e.message == 'command does not match the regex')));
     });
   });
+
+  group('A group of tests related to otp verb', () {
+    test('A test to verify otp verb for get operation', () {
+      String command = 'otp:get\n';
+      var enrollVerbParams =
+          VerbUtil.getVerbParam(VerbSyntax.otp, command.trim())!;
+      expect(enrollVerbParams['operation'], 'get');
+    });
+
+    test('A test to verify otp verb for validate operation', () {
+      String command = 'otp:validate:ABC123\n';
+      var enrollVerbParams =
+          VerbUtil.getVerbParam(VerbSyntax.otp, command.trim());
+      expect(enrollVerbParams!['operation'], 'validate');
+      expect(enrollVerbParams['otp'], 'ABC123');
+    });
+  });
 }
 
 Map getVerbParams(String regex, String command) {